requirements.txt
Flask==3.1.2
PyPDF2==3.0.1
python-docx==1.2.0
beautifulsoup4==4.14.3
striprtf==0.0.29
odfpy==1.4.1
